pyproject.toml
[project]
name = "laplacelab-remote-access"
version = "0.1.0"
description = "A FastAPI + Vite project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.1",
    "gunicorn>=23.0.0",
    "httpx>=0.27.2",
    "loguru>=0.7.3",
    "nanoid>=2.0.0",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.9.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
